pyproject.toml
[project]
name = "mycaddy"
version = "0.1.0"
description = "💻Personal Caddy server for me"
readme = "README.md"
authors = [{ name = "Dowon", email = "ks2515@naver.com" }]
requires-python = ">=3.10"
license = "MIT"
classifiers = [
    "Programming Language :: Go",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Server",
]
keywords = ["caddy", "wsgi", "asgi", "web server", "http server"]
dependencies = []

[dependency-groups]
dev = [
    "pdm-backend",
    "ruff",
    "pre-commit-uv",
    "tomli",
    "pbs-installer[all]",
    "starlette",
]

[build-system]
requires = [
    "pdm-backend",
    "tomli; python_version < '3.11'",
    "pbs-installer[all]",
]
build-backend = "pdm.backend"

[tool.uv]
default-groups = "all"

[tool.pdm.build]
is-purelib = false
excludes = ["bin/*", ".gitignore", "*.py"]

[tool.pdm.build.wheel-data]
scripts = ["${BUILD_DIR}/bin/*"]

[tool.ruff.lint]
select = [
    "A",
    "B",
    "C4",
    "C90",
    "E",
    "EM",
    "F",
    "FA",
    "FBT",
    "FURB",
    "I001",
    "ISC",
    "N",
    "PD",
    "PERF",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TC",
    "TRY",
    "UP",
    "W",
]
ignore = ["B905", "E501"]
unfixable = ["F401"]

[tool.cibuildwheel]
archs = "auto64"
build-frontend = "build[uv]"
test-sources = ["app.py", "Caddyfile"]
test-command = [
    "mycaddy --version",
    "mycaddy start --config Caddyfile",
    "uvx --from xh-bin xh https://localhost:8000",
    "uvx --from xh-bin xh https://localhost:8001",
    "mycaddy stop",
]

[tool.cibuildwheel.linux]
before-build = [
    "dnf install -y wget",
    "wget https://git.io/go-installer.sh && bash go-installer.sh",
    "source /root/.bashrc",
]

[tool.cibuildwheel.linux.environment]
PATH = "/root/.go/bin:$PATH"
